Return long instead of int when we are counting countries by name.
Should be in f9a8d41797afa7b6678bf7d8c44fb6c0894a2761 commit.

No functional changes.

+++ src/main/java/ru/mystamps/web/dao/CountryDao.java
@@ -25,7 +25,7 @@ import ru.mystamps.web.entity.Country;
 import ru.mystamps.web.service.dto.SelectEntityDto;
 
 public interface CountryDao extends PagingAndSortingRepository<Country, Integer> {
-	int countByName(String name);
+	long countByName(String name);
 	int countByNameRu(String name);
 	
 	@Query(
